feat: add metadata around pages
