+++ src/app/(auth)/register/page.tsx
@@ -1,17 +1,13 @@
-'use client'
-
 import { AuthHero } from '@/components/auth/auth-hero'
 import { RegisterForm } from '@/components/auth/register-form'
-import { useIsMounted } from '@/hooks/useIsMounted'
-import { useEffect } from 'react'
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+	title: 'PoliEats - Register',
+	description: 'Register page for PoliEats',
+}
 
 export default function Register() {
-	const isMounted = useIsMounted()
-
-	useEffect(() => {
-		isMounted()
-	}, [isMounted])
-
 	return (
 		<div className="flex flex-col items-center justify-center h-screen p-4 md:p-0 max-w-md mx-auto">
 			<AuthHero />
