+++ src/app/page.tsx
@@ -1,7 +1,10 @@
-'use client'
-
 import { Navbar } from '@/components/navbar'
 import { ProfileDialog } from '@/components/profile-dialog'
+
+export const metadata = {
+	title: 'PoliEats - Home',
+	description: 'Home page for PoliEats',
+}
 
 export default function Home() {
 	return (
